telemetry: don't record net i/o loop spans in scuba (#1908)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1908

These are used primarily to provide context for any logged events, and should not be recorded as separate spans in scuba.

We introduce a general mechanism, exposing the key `hyperactor_telemetry::skip_record` which, if included in a span, causes the downstream scuba subscribe to skip recording the span as its own discrete event. This also prevents us from entering and exiting the spans accordingly.

We do maintain the recorded fields, which provide important context; they continue to propagate to the individually recorded entries.
ghstack-source-id: 326152711
exported-using-ghexport

Reviewed By: vidhyav

Differential Revision: D87273700

fbshipit-source-id: 37ef580b0eea0226964c9d1539dc64ea1c4c3e31