Add torch-requirements file for the nightly version (#2025)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2025

Fixes https://github.com/meta-pytorch/monarch/issues/1978

Separate out the torch requirement into a different file, since it requires a different
index. We don't want `build-requirements.txt` to overwrite it.

It is not used in CI because we use "matrix.torch-spec" to change the torch version we
download. We should ensure the requirements file always matches the tested version.

Reviewed By: allenwang28

Differential Revision: D87802441

fbshipit-source-id: 8031aff8b8416e8cc67f3c1ea7e7ff4d181fb13a