clarify parsing of names and identifiers; make `Name` concrete syntax unambiguous (#2038)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2038

A previous change, D83495465, changed the `Name` format to separate identifiers and uuids with a '_' character. This introduces a parsing ambiguity, because we allow '_' to appear in names, and thus we can no longer unambiguously distinguish between a `Reserved` and `Suffixed` name.  We ran into this in practice because we had actors that used '_' characters in their names.

This change makes parsing lawful: to make the `Name` representation unambiguous, we reintroduce the '-' separator, and permit it as a name *element*. We retain the previous definition of an *identifier*, and apply this check to names that are constructed via `Name::new` and `Name::new_reserved`, making these functions fallible.
ghstack-source-id: 327072952
exported-using-ghexport

Reviewed By: dulinriley

Differential Revision: D88277771

fbshipit-source-id: d6607c7bfe53aba898429e481ccf12238a152147