Remove torch-op special call path in tensor engine (#1990)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1990

This is part of the work to remove a direct libtorch dependency. We now just route through python. A few months ago we tested that this is not slower because of so much wrapping/unwrapping to python objects anyway. We can re-enable faster execution in the future if needed.
ghstack-source-id: 326999499
exported-using-ghexport

Reviewed By: mariusae, dulinriley

Differential Revision: D87807456

fbshipit-source-id: a6f99e748e798735ed195fcb010ed49caf472802