Make SupervisionError::new_err return the correct error type (#2072)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2072

The `new_err` method is supposed to create a new PyErr with the same type.
This was accidentally set to the supertype RuntimeError. This should be `SupervisionError`
instead. Else `PyErr::new` and `SupervisionError::new_err` give slightly different exceptions.

Also follow this pattern for other exceptions in the same file to avoid accidental copy-paste.

Fixed and changed to use this shorthand method wherever we can.

Reviewed By: mariusae

Differential Revision: D88436495

fbshipit-source-id: 0ac652c8c7d975b33191c20230aac6ec95740f1c