better patches for vectors and hash maps; add option patch, error cases (#2078)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2078

In this change we clean up and clarify patches for vectors and hash maps. Patch is made a fallible operation, and we do not accept patches on incompatible values. We do not attempt to make `patch` a transactional operation.

This change also introduces a patch for `Option<T>`.
ghstack-source-id: 327883129
exported-using-ghexport

Reviewed By: dulinriley

Differential Revision: D88523020

fbshipit-source-id: b3cb4078540e9d666213f2a8104b52dcb5c8fa8d