Fix otel teardown bug during unit tests (#2095)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2095

The root cause is that the test creates real Counter objects
that use the global OpenTelemetry meter. When Python shuts
down, the order of destruction between Python objects and the Rust
OpenTelemetry global state can cause a use-after-free or double-free.

The fix should be to mock the inner Rust objects in the tests that don't need
real telemetry,

Reviewed By: amirafzali

Differential Revision: D88752324

fbshipit-source-id: 1df0297590e6683eb4d756a415dc639f07a6ca35