Change ProcMeshAgent timeout from Stopped to Failed (#2081)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2081

Part of: https://github.com/meta-pytorch/monarch/issues/2027

When a process goes down and we cannot reach the ProcMeshAgent, we used to use
the "Stopped" actor status. But Stopped is handled differently than Failed, Stopped is considered
a user-requested stop, and that the actor stopped cleanly. A process exit is not that, and
should follow the error propagation path up to the client.

Also, with Failed we can include a more helpful message saying the process was
crashed, which should help troubleshooting.

Reviewed By: mariusae

Differential Revision: D88529146

fbshipit-source-id: 188235cb46e8f4666bfab9d538bcb6bcdeb469bb