Fix deadlock in intern_typename (#2112)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2112

There can be a deadlock in DashMap::entry if it is called while any other references are alive.
To avoid this, separate the query for live keys from the insertion path, so the reference is
not alive. There's a small performance cost on the first insert, but the read path still only
has a single query.
Should be negligible, and helps avoid deadlocks.

Fixes hyperactor_named tests::test_ports

Reviewed By: shayne-fletcher, mariusae

Differential Revision: D88891169

fbshipit-source-id: 11aefc4cec1c3eddf9ea712f8a302d6f635a1ee4