Add requirement for Handler<SupervisionFailureMessage> to ProcMesh::spawn (#2126)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2126

Part of: https://github.com/meta-pytorch/monarch/issues/2027

In order for all actor types to support supervising the actors they own, we need to add
a capability to spawn, which requires the context to handle supervision messages.

Add this trait bound to ProcMesh::spawn, and further that to all callers.
A lot of tests used `Instance<()>` to spawn actors, so this was changed to
`Instance<TestRootClient>` to support these messages. The root client just
panics if it actually receives one of these messages.

The `global_root_client` used for non-tests now supports this message via its type,
but it doesn't actually have a message handling loop for the message. That should be ok,
as this is only used for v0.

Reviewed By: mariusae

Differential Revision: D88877061

fbshipit-source-id: 61f58fb9cb853d4c3683fc1a83c8938a4e1a649b