Fix serialize and deserialize of ValueMesh with bincode (#2145)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2145

ValueMesh was not serializable with bincode because it tried to use a self-describing format
with a serde tag, and bincode doesn't work with deserialize_any.

Remove this tag, and add tests that bincode serialization works. This is the serialization
we use with Mailbox, so it's important that it is supported.

While the JSON size may increase by a little bit, bincode is much smaller and the only format
we really use seriously, so I think this is worthwhile.

Reviewed By: shayne-fletcher

Differential Revision: D89214898

fbshipit-source-id: dd3eceae1d28b74c84a8dc0020cb4645d4ffb99f