Add Channel::Alias (#2141)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2141

In the Lightning integration, there is a need that we want to explicitly to config the client and worker hosts to:

1. use public IP address and an explicit port as the `Host`'s frontend address;
2. bind the port to INADDR_ANY, since AWS does not allow the port being bound to the public IP address.

In order to enable this configuration, this diff adds a new variant, `Alias` to channel. This variant provides 2 fields:

1. dial_to, which can be used by Lightning to set as the public IP address.
2. bind_to, which is used to specify how to bind the port.

It is Lightning's responsibility to ensure the network is configured in a way that the packages sent to `dial_to` would be routed to `bind_to`.

Reviewed By: mariusae

Differential Revision: D89190085

fbshipit-source-id: 5ef11d85b086a1cf9a75d92353e9f3e888262dd0