Add a reason to channel send error (#2188)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2188

Part of: https://github.com/meta-pytorch/monarch/issues/2041

On the NetTx sending pathway, there are some errors that would cause the channel to close,
and they mostly have an explainable reason for doing so. This wasn't making it back to the
errors stored in the envelope though.

Add the reason to SendError, and include this information in the whole set of queued messages
that are returned. Note that a reason for a single message is put on all the returned messages,
because all of them will fail for a reason possibly unrelated to themselves.

Reviewed By: shayne-fletcher, pzhan9

Differential Revision: D89562505

fbshipit-source-id: 2e76c4e38ff9a3f57e42baf8baeae8a41f182de9