Bind the undeliverable port on the global PythonActor (#2195)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2195

Fixes: https://github.com/meta-pytorch/monarch/issues/2041

Undeliverable messages sent from the root python client were not making it back to that
client, and were getting silently abandoned. This was because the handler for undeliverable
wasn't bound. Use a `.bind()` on the ActorHandle<PythonActor> to get the port bound.

Furthermore, there were some issues with how handle_supervision_event worked on the global
client. It wasn't called for immediate errors in its own message handlers, only for child supervision
events. This prevented us from calling `__supervise__` after an undeliverable message resulted
in an error.
Secondly, if a supervision event is not handled by the global client PythonActor, it was going
to ProcMeshAgent, which by default records events without crashing if it has a coordinator port
present.
For the global root client, we want to fail fast and loud, so we use process::exit after logging instead.
This in general should never happen, because `__supervise__` always either (a)
handles the event, or (b) causes a crash anyways. But to be defensive we never want to
silently drop these events.

Reviewed By: shayne-fletcher

Differential Revision: D89562504

fbshipit-source-id: eaeba4edec607407e107b54456b7cdf83a2adff0