Add Kubernetes support via SkyPilot (#2083)
Summary:
Hi Monarch community 👋

This PR adds support for running Monarch workloads on Kubernetes and cloud VMs via [SkyPilot](https://github.com/skypilot-org/skypilot).

Adds a new `SkyPilotJob` class that provisions Monarch workers on K8s (or any of the 20+ cloud providers supported by SkyPilot):

```python
job = SkyPilotJob(
    meshes={"workers": 2},
    resources=sky.Resources(cloud=sky.Kubernetes(), accelerators="H200:1"), # SkyPilot resources object https://skypilot.readthedocs.io/en/latest/reference/yaml-spec.html#resources
    cluster_name="my-cluster",
    file_mounts={"/local/path": "/worker/path"}, # For any code/data sync
)
```

With this I was able to run the getting started example from Monarch docs on my multi-node Coreweave Kubernetes cluster with H200s. Full example in `examples/skypilot_getting_started.py`.

**Notes:**
- **Cold start:** Currently builds Monarch from source on workers to keep client/worker versions in sync (~7-10 min). A pre-built container image would reduce this to ~30 seconds.
- **Networking:** Assumes workers are in the same network as the client. If client-worker communication is not very chatty, SkyPilot's SSH tunneling could remove this requirement.

 ---

This is an early PR - I'm looking for feedback on the direction. I'm a maintainer of the SkyPilot project and was trying to run Monarch on my k8s cluster. Figured an integration with SkyPilot would be a good fit.

Happy to iterate on the implementation if this aligns with where you'd like Monarch to go.

Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2083

Reviewed By: ahmadsharif1

Differential Revision: D88994552

Pulled By: colin2328

fbshipit-source-id: a89b6acb02218af525ab23ffa131b60687aab84d