Refactor actor init message (#2205)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2205

This makes it more clear that the actor init message can be creating at the same time as the call to allocate the python actor by refactoring ActorMesh._create. It gets inlined into the ProcMesh spawn code, and calls a method that returns the PythonMessage that is the init message.

However, we currently _cannot_ remove the actual cast of the init message. This is because actor init itself uses the context of the init message to initialize its rank. This information is not available to the rust actor's init method so we cannot handle our init method from within it.

Nevertheless, this change makes it clearer how to merge the init when it is possible.
ghstack-source-id: 331002976

Reviewed By: mariusae

Differential Revision: D89700862

fbshipit-source-id: 109e54acd29b7e304e2773a0af0c6dff611ad5d0