Fix a race in tensor engine (#2200)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2200

It is possible for an actor message to try to lookup a broker actor before the mesh controller has actually launched it. I started to see flaky tests from this recently. This is a workaround to stop the tests from being flaky because we know that the broker was spawned and we just have to wait for it to come into existence. The better solution, which requires much more refactoring, is to have the controller to spawn brokers and pass them as an argument to the controller so that it also has the broker references.
ghstack-source-id: 331002974
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D89690423

fbshipit-source-id: 6edcbf877f3303a27fc8273d06706cb865b5f180