Fix test_job and test_python_actors (#2214)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2214

Part of: https://github.com/meta-pytorch/monarch/issues/2213

test_job.py tests have flaky failures, described in the issue. This doesn't fix them, but
skips them in Github CI. We'll follow up to fix them when we have time.

test_proc_mesh_stop_after_actor_mesh_stop is not part of that task, but also had an
occasional failure due to a race condition between the actor constructor and stop. The stop
is sent to a different actor (ProcMeshAgent) and doesn't have ordering guarantees with
messages to the actor itself.

Reviewed By: vidhyav

Differential Revision: D89894540

fbshipit-source-id: ded37b78eea8ff09c15e1dc46b5eb27f7ed5b294